create distribution module
