+++ pyrenew/distributions/__init__.py
@@ -1,0 +1,7 @@
+# numpydoc ignore=GL08
+
+from pyrenew.distributions.censorednormal import CensoredNormal
+
+__all__ = [
+    "CensoredNormal",
+]
